refector: employ organize package and others
