+++ src/main/java/com/rgv04/hr/api/controller/JobHistoryController.java
@@ -1,4 +1,4 @@
-package com.rgv04.hr.controller;
+package com.rgv04.hr.api.controller;
 
 import java.util.List;
