+++ src/main/java/com/rgv04/hr/core/HrLinks.java
@@ -6,7 +6,7 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 import org.springframework.hateoas.Link;
 import org.springframework.stereotype.Component;
 
-import com.rgv04.hr.controller.JobHistoryController;
+import com.rgv04.hr.api.controller.JobHistoryController;
 
 @Component
 public class HrLinks {
