+++ src/main/java/com/rgv04/hr/domain/assembler/EmployeeAssembler.java
@@ -1,4 +1,4 @@
-package com.rgv04.hr.domain.employee;
+package com.rgv04.hr.domain.assembler;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,7 +6,10 @@ import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.server.mvc.RepresentationModelAssemblerSupport;
 import org.springframework.stereotype.Component;
 
-import com.rgv04.hr.controller.EmployeeController;
+import com.rgv04.hr.api.controller.EmployeeController;
+import com.rgv04.hr.domain.dto.EmployeeManagerModel;
+import com.rgv04.hr.domain.dto.EmployeeModel;
+import com.rgv04.hr.domain.model.Employee;
 
 @Component
 public class EmployeeAssembler extends RepresentationModelAssemblerSupport<Employee, EmployeeModel> {
