+++ src/main/java/com/rgv04/hr/domain/dto/EmployeeOnlyFirstNameAndHireDate.java
@@ -1,4 +1,4 @@
-package com.rgv04.hr.domain.employee;
+package com.rgv04.hr.domain.dto;
 
 import java.time.OffsetDateTime;
