+++ src/main/java/com/rgv04/hr/domain/dto/EmployeeSumary.java
@@ -1,4 +1,4 @@
-package com.rgv04.hr.domain.employee;
+package com.rgv04.hr.domain.dto;
 
 public interface EmployeeSumary {
