+++ src/main/java/com/rgv04/hr/domain/jobhistory/JobHistoryModelAssembler.java
@@ -6,7 +6,7 @@ import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.server.mvc.RepresentationModelAssemblerSupport;
 import org.springframework.stereotype.Component;
 
-import com.rgv04.hr.controller.JobHistoryController;
+import com.rgv04.hr.api.controller.JobHistoryController;
 import com.rgv04.hr.core.HrLinks;
 
 @Component
