+++ src/main/java/com/rgv04/hr/domain/region/assembler/RegionAssembler.java
@@ -6,7 +6,7 @@ import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.server.mvc.RepresentationModelAssemblerSupport;
 import org.springframework.stereotype.Component;
 
-import com.rgv04.hr.controller.RegionController;
+import com.rgv04.hr.api.controller.RegionController;
 import com.rgv04.hr.domain.region.Region;
 import com.rgv04.hr.domain.region.assembler.model.RegionModel;
