+++ src/main/java/com/rgv04/hr/domain/repository/EmployeeRepository.java
@@ -1,10 +1,13 @@
-package com.rgv04.hr.domain.employee;
+package com.rgv04.hr.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+
+import com.rgv04.hr.domain.dto.EmployeeSumary;
+import com.rgv04.hr.domain.model.Employee;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Long>, EmployeeRepositoryQueries {
