+++ src/main/java/com/rgv04/hr/domain/repository/EmployeeRepositoryQueries.java
@@ -1,6 +1,9 @@
-package com.rgv04.hr.domain.employee;
+package com.rgv04.hr.domain.repository;
 
 import java.util.List;
+
+import com.rgv04.hr.domain.dto.EmployeeFilter;
+import com.rgv04.hr.domain.dto.EmployeeOnlyFirstNameAndHireDate;
 
 public interface EmployeeRepositoryQueries {
